--- src/pages/Main/Carousel 2.js
@@ -1,15 +1,0 @@
-const Carousel = () => {
-  //key, imageURL, linkPath
-
-  return (
-    <div className="carousel">
-      <button className="prevImage" />
-      <div className="slidingImages">
-        <div className="slidingBar" />
-      </div>
-      <button className="nextImage" />
-    </div>
-  );
-};
-
-export default Carousel;
