+++ src/pages/Main/Pagination.js
@@ -2,6 +2,30 @@ import React from 'react';
 import ItemList from './ItemList';
 import './pagination.scss';
 
+/** 
+const [products, setProducts] = useState([]);
+const [currentPage, setCurrentPage] = useState(1);
+const [searchParams, setSearchParams] = useSearchParams(`?page=1&pageSize=9`);
+const page = searchParams.get('page');
+console.log('main', searchParams);
+
+useEffect(() => {
+  fetch(`http://10.58.7.174:3000/?page=${page}`, {
+    method: 'GET',
+  })
+    .then(res => res.json())
+    .then(data => setProducts(data.data));
+}, [page]);
+
+<Pagination
+  products={products}
+  currentPage={currentPage}
+  changeCurrentPage={setCurrentPage}
+  paramsSave={setSearchParams}
+/>;
+
+*/
+
 const Pagination = ({
   products,
   changeCurrentPage,
