--- src/pages/Main/card 2.scss
@@ -1,31 +1,0 @@
-.card {
-  margin-bottom: 70px;
-  .thumbNailImage {
-    width: 381px;
-    aspect-ratio: 1/1;
-    object-fit: fill;
-    margin-bottom: 10px;
-  }
-  .productName {
-    font-size: 20px;
-    margin-bottom: 10px;
-  }
-  .currentPrice {
-    font-size: 26px;
-    font-weight: 900;
-    margin-bottom: 10px;
-  }
-  .price {
-    font-size: 20px;
-    text-decoration: line-through;
-    margin-bottom: 10px;
-  }
-  .tagSpace {
-    margin-bottom: 10px;
-  }
-  .reviewCount {
-    font-size: 16px;
-    font-weight: 700;
-    margin-bottom: 10px;
-  }
-}
