--- src/pages/Main/itemlist 2.scss
@@ -1,5 +1,0 @@
-.itemList {
-  display: flex;
-  justify-content: space-around;
-  flex-wrap: wrap;
-}
