--- src/pages/Main/productsinmain 2.scss
@@ -1,12 +1,0 @@
-.titleMain {
-  text-align: center;
-  .bigTitle {
-    padding-bottom: 4px;
-    font-size: 30px;
-    font-weight: bold;
-  }
-  .description {
-    padding-bottom: 55px;
-    font-size: 24px;
-  }
-}
